+++ clients/apps/web/src/components/CustomerPortal/CustomerPortalTeam.tsx
@@ -7,8 +7,7 @@ import {
   useRemoveCustomerPortalMember,
   useUpdateCustomerPortalMember,
 } from '@/hooks/queries/customerPortal'
-import { useSafeCopy } from '@/hooks/clipboard'
-import { CONFIG } from '@/utils/config'
+import { useCopyMemberLoginLink } from '@/hooks/useCopyMemberLoginLink'
 import { validateEmail } from '@/utils/validation'
 import MoreVertOutlined from '@mui/icons-material/MoreVertOutlined'
 import { Client, schemas } from '@polar-sh/client'
@@ -61,7 +60,7 @@ export const CustomerPortalTeamSection = ({
   const updateMember = useUpdateCustomerPortalMember(api)
   const removeMember = useRemoveCustomerPortalMember(api)
   const addMember = useAddCustomerPortalMember(api)
-  const safeCopy = useSafeCopy(toast)
+  const copyMemberLoginLink = useCopyMemberLoginLink(organizationSlug)
 
   const [memberToRemove, setMemberToRemove] = useState<string | null>(null)
   const [loadingMembers, setLoadingMembers] = useState<Set<string>>(new Set())
@@ -140,18 +139,6 @@ export const CustomerPortalTeamSection = ({
         return next
       })
     }
-  }
-
-  const handleCopyLoginLink = async (memberEmail: string) => {
-    const link = `${CONFIG.FRONTEND_BASE_URL}/${organizationSlug}/portal/request?email=${encodeURIComponent(
-      memberEmail,
-    )}`
-    await safeCopy(link)
-    toast({
-      title: 'Login link copied',
-      description:
-        'Share it with the member so they can sign in to the portal with their email.',
-    })
   }
 
   const handleRemoveMember = async (memberId: string) => {
@@ -291,7 +278,7 @@ export const CustomerPortalTeamSection = ({
                           </DropdownMenuTrigger>
                           <DropdownMenuContent align="end">
                             <DropdownMenuItem
-                              onClick={() => handleCopyLoginLink(member.email)}
+                              onClick={() => copyMemberLoginLink(member.email)}
                               disabled={isLoading}
                             >
                               Copy login link
